split the set-based coverage into its own files
